+++ tests/src/lib/index.ts
@@ -1,0 +1,8 @@
+import '@lit-labs/ssr/lib/install-global-dom-shim';
+import { runScenario  } from "@holochain/tryorama";
+import concurrent from './concurrent.js';
+import test from "tape-promise/tape.js";
+
+test("syn concurrent test", async (t) => {
+  await runScenario(concurrent(t))
+})
